pyproject.toml
[tool.poetry]
name = "kolenka-backend"
version = "0.3.4"
description = ""
authors = ["Pavel <useless.dinosaur@veloc1.me>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.1"
peewee = "^3.13.1"
psycopg2-binary = "^2.8.4"
requests = "^2.23.0"
flask-cors = "^3.0.8"
bleach = "^3.1.1"
python-magic = "^0.4.15"
python-dotenv = "^0.12.0"
black = "^19.10b0"

[tool.poetry.dev-dependencies]
pytest = "^5.4.1"
pytest-cov = "^2.8.1"
black = "^19.10b0"
pytest-mock = "^2.0.0"
flake8 = "^3.8.3"
isort = "^5.6.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
